requirements.txt
streamlit
PyPDF2
scikit-learn
nltk
google-generativeai
reportlab
